Rename constant pool entries in place
The current use of ClassWriter(ClassReader,int) meant that all existing
constant pool entries from the ClassReader are copied as-is to the new
class, and any renamed constant pool entries are merely appended to the
pool rather than replacing the original non-renamed entries (as
specified in the javadoc of said ClassWriter constructor).

To avoid the waste of constant pool space by those unused non-renamed
entries, RenamingTransformer now creates a fresh ClassWriter (not
passing in the ClassReader).

Note that ASM does some reordering of the constant pool (and some other
attributes) that deviate from the order outputted by javac.