Fix ParameterAnnotationFixer on modern Minecraft
ProGuard has been updated since the inception of this fix in https://github.com/ModCoderPack/MCInjector/pull/3 and no longer has the original bug. Mojang also appears to be using newer versions of ProGuard where the bug is not present.
The fixer is now properly programmed to change nothing if it doesn't detect the broken PG behavior.
The fixer is now also optimized to use the visitor system instead of populating a ClassNode; it has been rewritten to do so.